// frontend/src/types/doctors.ts
export interface Doctor {
  id: string;
  name: string;
  description: string;
  profileUrl: string;
  phone: string;
  isVerified: boolean;
  email: string;
  address: string;
  lat: number;
  long: number;
  price: number;
  startDay: string;
  endDay: string;
  startTime: string;
  endTime: string;
  imagePath: string;
  coverImagePath: string;
}
export interface CBDShop {
  id: string;
  name: string;
  description: string;
  profileUrl: string;
  phone: string;
  isVerified: boolean;
  email: string;
  address: string;
  lat: number;
  long: number;
  price: number;
  startDay: string;
  endDay: string;
  startTime: string;
  endTime: string;
  imagePath: string;
  coverImagePath: string;
}

export interface HeadShop {
  id: string;
  name: string;
  description: string;
  profileUrl: string;
  phone: string;
  isVerified: boolean;
  email: string;
  address: string;
  lat: number;
  long: number;
  price: number;
  startDay: string;
  endDay: string;
  startTime: string;
  endTime: string;
  imagePath: string;
  coverImagePath: string;
}

export interface GrowEquipment {
  id: string;
  name: string;
  description: string;
  profileUrl: string;
  phone: string;
  isVerified: boolean;
  email: string;
  address: string;
  lat: number;
  long: number;
  price: number;
  startDay: string;
  endDay: string;
  startTime: string;
  endTime: string;
  imagePath: string;
  coverImagePath: string;
}
export interface Pharmacy {
  id: string;
  name: string;
  description: string;
  profileUrl: string;
  phone: string;
  isVerified: boolean;
  email: string;
  address: string;
  lat: number;
  long: number;
  price: number;
  startDay: string;
  endDay: string;
  startTime: string;
  endTime: string;
  imagePath: string;
  coverImagePath: string;
}